fix: surface BLE connect failures, and stop enums logging as garbage
A user report of "Android can't connect on app v3.3.0" root-caused to a bond
mismatch, not an app regression: the board's settings partition had been wiped
(HIL destructive tier / appcfg erase) while the phone kept its bond, so every
connect died in ~370 ms with BT_SECURITY_ERR_PIN_OR_KEY_MISSING before discovery
started. Reproduced live on a Pixel 9 Pro with the v3.3.0 release build; clearing
the stale bond fixed it (verified CONNECTED / L4 / ATT MTU 498).

Two defects made that far more expensive to diagnose than it should have been:

app: a failed connect was completely silent. connect() resolves false and the
press handler simply doesn't navigate, so every failure looked like a dead
button. Device rows now show a tappable failure indicator whose alert carries the
underlying reason plus the forget-and-re-pair recovery. Deliberately NOT
classified by error code: describeWriteError only asserts androidErrorCode 252
because that was hardware-measured (issue #92), and the connect-side ble-plx shape
has not been captured, so the reason is surfaced verbatim and the hint is phrased
as guidance rather than a diagnosis.

fw: security levels printed as garbage - "Failed to reach required security level
59609092, got 59609089" instead of 4 and 1. Under CONFIG_LOG_MODE_DEFERRED
arguments are packed, not passed as varargs, so no integer promotion happens;
cbprintf_cxx.h has promoting overloads for char/short/etc but none for enums, so
an enum takes the generic template, which reserves MAX(sizeof(T), sizeof(int))
bytes and then copies 4 bytes out of a 1-byte object - the true value survives in
the low byte, the upper 24 bits are adjacent stack. bt_security_t is
`typedef enum __packed`. This affects EVERY enum logged from C++ here, so the rule
is documented in fw/CLAUDE.md rather than only fixed at these six call sites.
Hardware-verified after flashing: "Peer security level changed to 4".

scripts: re-pair.py hardcoded the dev package, so pointing it at a release-build
bug silently drove the wrong app and reported "board not listed" - which blames
the board rather than the app. Adds --package/--release (default unchanged), a
fail-fast install check, and a startup line naming the package driven.

Verified: proto0 builds clean; firmware flashed and confirmed on hardware;
app tsc/lint clean, jest 546/546 (+9 new tests); re-pair --release re-paired and
verified CONNECTED / L4 / MTU 498.

+++ fw/src/bluetooth.cpp
@@ -500,10 +500,20 @@ static void security_changed(struct bt_conn *conn, bt_security_t level, enum bt_
 
     bt_addr_le_to_str(bt_conn_get_dst(conn), addr, sizeof(addr));
 
+    // (int) casts are REQUIRED, not cosmetic: bt_security_t is `typedef enum __packed`
+    // (1 byte) and this is a C++ TU, so Zephyr's log-argument packing takes the generic
+    // template in cbprintf_cxx.h - which reserves MAX(sizeof(T), sizeof(int)) = 4 bytes
+    // but then memcpy's 4 bytes out of the 1-byte object, pulling 3 bytes of adjacent
+    // stack into the upper bits. Without the cast this printed "level 66701313 err
+    // 66701314" (0x03F9C801 / 0x03F9C802) instead of "level 1 err 2" - the true value
+    // survives in the low byte only. The explicit char/short overloads in that header
+    // promote via `int tmp = arg + 0;`; enums have no such overload, so the promotion
+    // has to happen here. Applies to EVERY enum logged from C++ in this codebase - see
+    // fw/CLAUDE.md, "Logging an enum from C++ prints garbage without an (int) cast".
     if (!err) {
-        LOG_INF("Security changed: %s level %u", addr, level);
+        LOG_INF("Security changed: %s level %u", addr, (int)level);
     } else {
-        LOG_ERR("Security failed: %s level %u err %d", addr, level, err);
+        LOG_ERR("Security failed: %s level %u err %d", addr, (int)level, (int)err);
     }
 
     BtThreadCommand cmd;
@@ -573,7 +583,9 @@ static void pairing_failed(struct bt_conn *conn, enum bt_security_err reason) {
 
     bt_addr_le_to_str(bt_conn_get_dst(conn), addr, sizeof(addr));
 
-    LOG_INF("Pairing failed conn: %s, reason %d", addr, reason);
+    // (int): enum bt_security_err is 1 byte under -fshort-enums - see the cast rationale
+    // in security_changed() above.
+    LOG_INF("Pairing failed conn: %s, reason %d", addr, (int)reason);
 }
 
 static void bond_deleted(uint8_t id, const bt_addr_le_t *peer) {
@@ -813,7 +825,8 @@ void bt_state_connecting_handle_command(BtThreadContext *ctx, const BtThreadComm
 
         case BtThreadEvent::SECURITY_CHANGED:
             // What level of security did the peer upgrade to?
-            LOG_INF("Peer security level changed to %d", cmd->level);
+            // (int): bt_security_t is a 1-byte packed enum - see security_changed().
+            LOG_INF("Peer security level changed to %d", (int)cmd->level);
 
             if (cmd->level == REQUIRED_BT_SECURITY_LEVEL) {
                 LOG_DBG("Required security level achieved");
@@ -838,8 +851,11 @@ void bt_state_connecting_handle_command(BtThreadContext *ctx, const BtThreadComm
 
                 bt_state_change_to(ctx, BtThreadState::CONNECTED);
             } else {
+                // (int) on BOTH: REQUIRED_BT_SECURITY_LEVEL is BT_SECURITY_L4, which is
+                // itself a bt_security_t constant - so even the compile-time literal
+                // printed as 0x038D9004 instead of 4. See security_changed().
                 LOG_ERR("Failed to reach required security level %d, got %d instead",
-                        REQUIRED_BT_SECURITY_LEVEL, cmd->level);
+                        (int)REQUIRED_BT_SECURITY_LEVEL, (int)cmd->level);
 
                 // Disconnect from the peer (re-enabled for issue #232). With
                 // CONFIG_BT_SMP_SC_ONLY a below-L4 pairing can no longer *succeed*,
@@ -901,14 +917,15 @@ void bt_state_connected_handle_command(BtThreadContext *ctx, const BtThreadComma
 
         case BtThreadEvent::SECURITY_CHANGED:
             // What level of security did the peer upgrade to?
-            LOG_INF("Peer security level changed to %d", cmd->level);
+            // (int): bt_security_t is a 1-byte packed enum - see security_changed().
+            LOG_INF("Peer security level changed to %d", (int)cmd->level);
 
             if (cmd->level == REQUIRED_BT_SECURITY_LEVEL) {
                 LOG_DBG("Required security level achieved");
                 bt_state_change_to(ctx, BtThreadState::CONNECTED);
             } else {
-                LOG_WRN("Reached security %d, must reach %d to connect", cmd->level,
-                        REQUIRED_BT_SECURITY_LEVEL);
+                LOG_WRN("Reached security %d, must reach %d to connect", (int)cmd->level,
+                        (int)REQUIRED_BT_SECURITY_LEVEL);
                 // TODO: need to implement timeout to disconnect if we don't reach "CONNECTED" state
                 // within ~60s
             }
